dlgthingtalk: change how we compute the prediction from the state
Make the executed part of the state immutable, and always append
to the current history. Change "confirm" to be an enum: accepted,
proposed, confirmed.
