+++ lib/languages/dlgthingtalk/agent.js
@@ -9,8 +9,11 @@
 // See COPYING for details
 "use strict";
 
+const assert = require('assert');
 const ThingTalk = require('thingtalk');
 const Ast = ThingTalk.Ast;
+
+const { isExecutable } = require('./state_utils');
 
 /**
  * Run the dialogue, executing ThingTalk and invoking the policy at the
@@ -23,38 +26,6 @@ class DialogueAgent {
     constructor(executor, options) {
         this._rng = options.rng;
         this._executor = executor;
-    }
-
-    _isExecutable(stmt, confirm) {
-        let hasUndefined = false;
-        let needsConfirm = false;
-        if (stmt.isRule)
-            needsConfirm = true;
-        if (needsConfirm && !confirm)
-            return false;
-        const visitor = new class extends Ast.NodeVisitor {
-            visitValue(value) {
-                if (value.isUndefined)
-                    hasUndefined = true;
-                return true;
-            }
-
-            visitInvocation(value) {
-                if (value.schema.annotations.confirm)
-                    needsConfirm = needsConfirm || value.schema.annotations.confirm.toJS();
-                else
-                    needsConfirm = needsConfirm || value.schema.functionType === 'action';
-                return true;
-            }
-        };
-        stmt.visit(visitor);
-
-        if (hasUndefined)
-            return false;
-        if (needsConfirm && !confirm)
-            return false;
-
-        return true;
     }
 
     /**
@@ -73,8 +44,9 @@ class DialogueAgent {
         for (let i = 0; i < clone.history.length; i++) {
             if (clone.history[i].results !== null)
                 continue;
-            if (!this._isExecutable(clone.history[i].stmt, clone.history[i].confirm))
+            if (clone.history[i].confirm !== 'confirmed')
                 continue;
+            assert(isExecutable(clone.history[i].stmt));
 
             if (!anyChange) {
                 clone = state.clone();
