+++ lib/languages/dlgthingtalk/index.js
@@ -17,7 +17,7 @@ const SchemaRetriever = ThingTalk.SchemaRetriever;
 
 const ThingTalkSimulator = require('./simulator');
 const DialogueAgent = require('./agent');
-const { computeNewState, computePrediction } = require('./state_utils');
+const { computeNewState, computePrediction, prepareContextForPrediction } = require('./state_utils');
 
 module.exports = {
     async parse(code, options) {
@@ -59,47 +59,41 @@ module.exports = {
         return [code, entities];
     },
 
-    computeNewState(oldState, prediction) {
-        return computeNewState(oldState, prediction);
+    computeNewState(oldState, prediction, forTarget) {
+        return computeNewState(oldState, prediction, forTarget);
     },
 
     /**
-     * Compute the information that the neural network interpreting the user input
-     * must predict, to compute new state.
+     * Compute the information that the neural network must predict to compute the new state
+     * in a turn.
      *
-     * This should return a string representation, roughly corresponding to the
-     * delta between `oldState` and `newState`. The string should be parseable with
-     * .parse().
-     * Neither `oldState` nor `newState` must be modified in-place.
+     * This applies to either the network interpreting the user input, the one controlling the
+     * dialogue policy.
      *
-     * @param {ThingTalk.Ast.DialogueState} oldState - the previous dialogue state, before the user speaks
-     * @param {ThingTalk.Ast.DialogueState} newState - the new state of the dialogue, after the user speaks
-     * @param {string[]} sentence - the utterance from the user
-     * @param {Object} entities - entities contained in the utterance from the user or the context
-     * @return {string[]} - the delta to predict, as a sequence of tokens
-     */
-    computeUserPrediction(oldState, newState, sentence, entities) {
-        const prediction = computePrediction(oldState, newState);
-        return ThingTalk.NNSyntax.toNN(prediction, sentence, entities, { typeAnnotations: false });
-    },
-
-    /**
-     * Compute the information that the neural network controlling the dialogue policy
-     * must predict, to compute new state.
-     *
-     * This should return a new state object, roughly corresponding to the
+     * This should return a new state, roughly corresponding to the
      * delta between `oldState` and `newState`.
      * Neither `oldState` nor `newState` must be modified in-place.
      *
-     * @param {ThingTalk.Ast.DialogueState} oldState - the previous dialogue state, before the system speaks
-     * @param {ThingTalk.Ast.DialogueState} newState - the new state of the dialogue, after the system speaks
-     * @param {string[]} sentence - the utterance from the agent
-     * @param {Object} entities - entities contained in the utterance from the agent or the context
+     * @param {ThingTalk.Ast.DialogueState} oldState - the previous dialogue state, before the turn
+     * @param {ThingTalk.Ast.DialogueState} newState - the new state of the dialogue, after the turn
+     * @param {string} forTarget - who is speaking now: either `user` or `agent`
+     */
+    computePrediction(oldState, newState, forTarget) {
+        return computePrediction(oldState, newState, forTarget);
+    },
+
+    /**
+     * Convert the prediction to a sequence of tokens to predict.
+     *
+     * @param {ThingTalk.Ast.DialogueState} oldState - the previous dialogue state, before the turn
+     * @param {ThingTalk.Ast.DialogueState} newState - the new state of the dialogue, after the turn
+     * @param {string[]} sentence - the utterance spoken at this turn
+     * @param {Object} entities - entities contained in the utterance or the context
+     * @param {string} forTarget - who is speaking now: either `user` or `agent`
      * @return {string[]} - the delta to predict, as a sequence of tokens
      */
-    computeAgentPrediction(oldState, newState, sentence, entities) {
-        const prediction = computePrediction(oldState, newState);
-        return ThingTalk.NNSyntax.toNN(prediction, sentence, entities, { allocateEntities: true, typeAnnotations: false });
+    serializePrediction(prediction, sentence, entities, forTarget) {
+        return ThingTalk.NNSyntax.toNN(prediction, sentence, entities, { allocateEntities: forTarget === 'agent', typeAnnotations: false });
     },
 
     createSimulator(options = {}) {
@@ -110,18 +104,6 @@ module.exports = {
     },
 
     prepareContextForPrediction(context, forTarget) {
-        const clone = context.clone();
-
-        for (let item of clone.history) {
-            if (item.results === null)
-                continue;
-
-            if (forTarget === 'user' && item.results.results.length > 1)
-                item.results.results.length = 1;
-            else if (item.results.results.length > 3)
-                item.results.results.length = 3;
-        }
-
-        return clone;
+        return prepareContextForPrediction(context, forTarget);
     },
 };
