+++ lib/languages/dlgthingtalk/state_utils.js
@@ -14,87 +14,146 @@ const assert = require('assert');
 const ThingTalk = require('thingtalk');
 const Ast = ThingTalk.Ast;
 
-function computePrediction(oldState, newState) {
+function isExecutable(stmt) {
+    let hasUndefined = false;
+    const visitor = new class extends Ast.NodeVisitor {
+        visitValue(value) {
+            if (value.isUndefined)
+                hasUndefined = true;
+            return true;
+        }
+    };
+    stmt.visit(visitor);
+    return !hasUndefined;
+}
+
+function shouldAutoConfirmStatement(stmt) {
+    if (stmt.isRule)
+        return false;
+
+    let needsConfirm = false;
+    const visitor = new class extends Ast.NodeVisitor {
+        visitInvocation(value) {
+            if (value.schema.annotations.confirm)
+                needsConfirm = needsConfirm || value.schema.annotations.confirm.toJS();
+            else
+                needsConfirm = needsConfirm || value.schema.functionType === 'action';
+            return true;
+        }
+    };
+    stmt.visit(visitor);
+    return !needsConfirm;
+}
+
+function computePrediction(oldState, newState, forTarget) {
     if (oldState === null)
         return newState;
 
-    // This assertion (and the one below) does not hold in case of
-    //
-    // $dialogue @org.thingpedia.dialogue.transaction.sys_propose_refined_query;
-    // now => (@uk.ac.cam.multiwoz.Restaurant.Restaurant()), (...) => notify
-    // #[results=[...]];
-    // #[count=108]
-    // now => (result(@uk.ac.cam.multiwoz.Restaurant.Restaurant[1])), id == "str:ENTITY_uk.ac.cam.multiwoz.Restaurant:Restaurant::9"^^uk.ac.cam.multiwoz.Restaurant:Restaurant => notify;
-    // now => @uk.ac.cam.multiwoz.Restaurant.make_reservation(restaurant=$?, book_time=$?, book_day=$?, book_people=$?);
-    //
-    // in that case, the previous state (the context) includes the first and third program
-    // so the first program in the prediction (the `result()` program) is not
-    // compatible with the action
-    //
-    // the next state (the user prediction) includes a modified first program and the third program
-    // it has length two, which is less than the length of the current state
-    //
-    //assert(newState.history.length >= oldState.history.length);
-
     const deltaState = new Ast.DialogueState(null, newState.policy, newState.dialogueAct, newState.dialogueActParam, []);
 
-    // walk forward through both states until we find a history item that is not equal
-    let i = 0;
-    while (i < Math.min(newState.history.length, oldState.history.length)) {
-        if (!oldState.history[i].equals(newState.history[i]))
-            break;
-        i++;
-    }
-    /*if (i < oldState.history.length)
-        assert(newState.history[i].compatible(oldState.history[i]));
-    */
+    // walk forward the state of oldState and newState
+    // until we find a program that is not confirmed in oldState
+    // the delta starts in newState at that position
 
-    deltaState.history = newState.history.slice(i);
-    for (let i = 0; i < deltaState.history.length; i++) {
-        if (deltaState.history[i].results !== null) {
-            console.log(oldState.prettyprint());
-            console.log(newState.prettyprint());
-        }
-        assert(deltaState.history[i].results === null);
+    let i;
+    for (i = 0; i < Math.min(oldState.history.length, newState.history.length); i++) {
+        const oldItem = oldState.history[i];
+        const newItem = newState.history[i];
+
+        if (oldItem.confirm !== 'confirmed')
+            break;
+
+        assert(oldItem.equals(newItem));
     }
+
+    // all new state items after i are part of the delta
+    deltaState.history = newState.history.slice(i).map((item) => item.clone());
+
+    // check that the results is null for everything in the prediction
+    // and reset confirm to false (the default) for everything autoconfirmable
+    for (let i = 0; i < deltaState.history.length; i++) {
+        assert(deltaState.history[i].results === null);
+
+        if (forTarget === 'user' && shouldAutoConfirmStatement(deltaState.history[i].stmt))
+            deltaState.history[i].confirm = 'accepted';
+    }
+
     return deltaState;
 }
 
-function computeNewState(state, prediction) {
-    if (state === null)
-        return prediction;
+function computeNewState(state, prediction, forTarget) {
+    const clone = new Ast.DialogueState(null, prediction.policy, prediction.dialogueAct, prediction.dialogueActParam, []);
 
-    // walk backwards through state until we find an history item that is compatible with
-    // the first program in the prediction, then override everything else afterwards
-
-    const clone = new Ast.DialogueState(null, prediction.policy, prediction.dialogueAct, prediction.dialogueActParam, state.history.slice());
-    if (prediction.history.length === 0)
-        return clone;
-
-    /*
-    console.log();
-    console.log('--- computeNewState');
-    console.log(clone.prettyprint());
-    console.log(prediction.prettyprint());
-    */
-
-    for (let i = clone.history.length-1; i >= 0; i--) {
-        const item = clone.history[i];
-
-        if (item.compatible(prediction.history[0])) {
-            if (i > 0)
-                clone.history = clone.history.slice(0, i).concat(prediction.history);
-            else
-                clone.history = prediction.history.slice();
-            return clone;
+    // append all history elements that were confirmed
+    if (state !== null) {
+        for (let oldItem of state.history) {
+            if (oldItem.confirm !== 'confirmed')
+                break;
+            clone.history.push(oldItem);
         }
     }
-    // if we did not find a compatible program, then append to the end
-    clone.history.push(...prediction.history);
+
+    const autoConfirm = forTarget === 'user';
+    // append the prediction items, and set the confirm bit if necessary
+    for (let newItem of prediction.history) {
+        let cloneItem = newItem;
+        if (cloneItem.confirm === 'accepted' && autoConfirm && isExecutable(cloneItem.stmt) && shouldAutoConfirmStatement(cloneItem.stmt)) {
+            cloneItem = cloneItem.clone();
+            cloneItem.confirm = 'confirmed';
+        }
+        clone.history.push(cloneItem);
+    }
+
+    return clone;
+}
+
+function prepareContextForPrediction(context, forTarget) {
+    const clone = new Ast.DialogueState(null, context.policy, context.dialogueAct, context.dialogueActParam, []);
+
+    // walk through all items in the history, find the last with results
+
+    let i, lastItem;
+    for (i = 0; i < context.history.length; i++) {
+        const item = context.history[i];
+        if (item.results === null)
+            break;
+        lastItem = item;
+    }
+
+    // add a copy of the last item with results, and trim the result list to 1 or 3
+    if (lastItem) {
+        const cloneItem = lastItem.clone();
+        if (forTarget === 'user' && cloneItem.results.results.length > 1)
+            cloneItem.results.results.length = 1;
+        else if (cloneItem.results.results.length > 3)
+            cloneItem.results.results.length = 3;
+        clone.history.push(lastItem);
+    }
+
+    // append the unconfirmed/unexecuted results
+    for (; i < context.history.length; i++) {
+        const item = context.history[i];
+        assert(item.results === null);
+
+        // about this assertion:
+        //
+        // on the agent side, we just executed the state, so all confirmed statements
+        // must have results
+        //
+        // on the user side, the agent just spoke; the agent never introduces confirmed statements,
+        // because statements must be confirmed by the user, so this assertion is also true
+        assert(item.confirm !== 'confirmed');
+
+        clone.history.push(item);
+    }
+
     return clone;
 }
 
 module.exports = {
+    isExecutable,
+
     computePrediction,
     computeNewState,
+    prepareContextForPrediction,
 };
