+++ lib/languages/multidst/index.js
@@ -163,34 +163,23 @@ module.exports = {
      * must predict, to compute new state.
      *
      * This should return a string representation, roughly corresponding to the
-     * delta between `oldState` and `newState`. The string should be parseable with
-     * .parse().
+     * delta between `oldState` and `newState`.
      * Neither `oldState` nor `newState` must be modified in-place.
      *
      * @param {any} oldState - the previous dialogue state, before the user speaks
      * @param {any} newState - the new state of the dialogue, after the user speaks
-     * @return {string[]} - the delta to predict, as a sequence of tokens
+     * @return {any} - the delta to predict
      */
-    computeUserPrediction(oldState, newState) {
+    computePrediction(oldState, newState, forTarget) {
         // always predict newState a-new
-        return newState.prettyprint().split(' ');
+        return newState;
     },
 
-    /**
-     * Compute the information that the neural network controlling the dialogue policy
-     * must predict, to compute new state.
-     *
-     * This should return a new state object, roughly corresponding to the
-     * delta between `oldState` and `newState`.
-     * Neither `oldState` nor `newState` must be modified in-place.
-     *
-     * @param {any} oldState - the previous dialogue state, before the system speaks
-     * @param {any} newState - the new state of the dialogue, after the system speaks
-     * @return {string[]} - the delta to predict, as a sequence of tokens
-     */
-    computeSystemPrediction(oldState, newState) {
-        // we don't care about the agent dialogue policy in this task
-        return [];
+    serializePrediction(prediction, sentence, entities, forTarget) {
+        if (forTarget === 'user')
+            return prediction.prettyprint().split(' ');
+        else // we don't care about the agent dialogue policy in this task
+            return [];
     },
 
     createSimulator() {
