+++ lib/sentence-generator/index.js
@@ -319,8 +319,11 @@ class MinibatchDialogueGenerator {
                 assert(agent);
                 assert(user);
 
-                agent_target = agentState.prettyprint();
-                user_target = userState.prettyprint();
+                const agentPrediction = this._target.computePrediction(dlg.context, agentState, 'agent');
+                agent_target = agentPrediction.prettyprint();
+
+                const userPrediction = this._target.computePrediction(agentState, userState, 'user');
+                user_target = userPrediction.prettyprint();
             }
 
             continuations.put(dlg, {
